+++ core-bundle/src/Resources/contao/forms/Form.php
@@ -89,8 +89,6 @@ class Form extends Hybrid
 
 	/**
 	 * Generate the form
-	 *
-	 * @return string
 	 */
 	protected function compile()
 	{