+++ core-bundle/src/Resources/contao/library/Contao/Automator.php
@@ -10,6 +10,7 @@
 
 namespace Contao;
 
+use Contao\CoreBundle\OptIn\OptIn;
 use Symfony\Component\Console\Input\ArgvInput;
 use Symfony\Component\Console\Output\NullOutput;
 
@@ -211,10 +212,26 @@ class Automator extends System
 			return;
 		}
 
-		foreach ($objMember as $objModel)
+		while ($objMember->next())
 		{
-			$objModel->delete();
+			$objMember->delete();
 		}
+
+		// Add a log entry
+		$this->log('Purged the unactivated member registrations', __METHOD__, TL_CRON);
+	}
+
+	/**
+	 * Purge opt-in tokens
+	 */
+	public function purgeOptInTokens()
+	{
+		/** @var OptIn $optIn */
+		$optIn = \System::getContainer()->get('contao.opt-in');
+		$optIn->purgeTokens();
+
+		// Add a log entry
+		$this->log('Purged the expired double opt-in tokens', __METHOD__, TL_CRON);
 	}
 
 	/**