+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## DEV
 
+ * Increase the back end preview image dimensions (see #246).
+ * Add the "contao.slug" service (see #222).
+ * Add the "contao.opt-in" service (see #196).
+ * Add the onshow_callback (see #235).
  * Enable drag and drop for templates (see #223).
  * Add the integrity attribute when loading jQuery from CDN (see contao/core-bundle#702).
  * Add methods to retrive past and upcoming dates in the event reader (see #175).